requirements.txt
alembic==1.18.4
amazon-transcribe==0.6.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
awscrt==0.26.1
bcrypt==3.2.0
boto3==1.42.59
botocore==1.42.59
brotli==1.2.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cssselect2==0.9.0
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.134.0
fonttools==4.61.1
fpdf2==2.8.7
greenlet==3.3.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
motor==3.7.1
numpy==2.4.2
packaging==26.0
pandas==3.0.1
passlib==1.7.4
pillow==12.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydyf==0.12.1
pymongo==4.16.0
pyparsing==3.3.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.47
starlette==0.52.1
tinycss2==1.5.1
tinyhtml5==2.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
weasyprint==68.1
webencodings==0.5.1
websockets==12.0
zopfli==0.4.1